added the checkstyle to pom and refactored the packages
