+++ samples/TestEntityExpansionLimit/src/main/java/xxe/jaxb/DemoWithSecurityManager.java
@@ -16,45 +16,51 @@
  *  under the License.
  */
 
-package xxe;
+package xxe.jaxb;
+
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
+import org.apache.xerces.util.SecurityManager;
 import org.xml.sax.InputSource;
 import org.xml.sax.SAXException;
 import org.xml.sax.XMLReader;
+
 import xxe.entity.Foo;
 
-import javax.xml.XMLConstants;
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
+import java.io.InputStreamReader;
+import java.io.UnsupportedEncodingException;
 import javax.xml.bind.JAXBContext;
 import javax.xml.bind.JAXBException;
 import javax.xml.bind.Unmarshaller;
 import javax.xml.parsers.ParserConfigurationException;
+import javax.xml.parsers.SAXParser;
 import javax.xml.parsers.SAXParserFactory;
 import javax.xml.transform.sax.SAXSource;
-import java.io.FileNotFoundException;
-import java.io.FileReader;
-
 
 
 /**
- * DemoWithoutSecurityManager without overiding the EntityExpansionLimit from xerces.util.SecurityManager
- * The default value for EntityExpansion limit from xerces.util.SecurityManager will be used
- * even an environment property is added for the EntityExpansionLimit.
+ * DemoWithSecurityManager with overiding the EntityExpansionLimit from xerces.util.SecurityManager
  */
 
-public class DemoWithoutSecurityManager {
-    private static final Log log = LogFactory.getLog(DemoWithoutSecurityManager.class);
-    private XMLReader xmlReader;
+public class DemoWithSecurityManager {
+    private static final Log log = LogFactory.getLog(DemoWithSecurityManager.class);
 
     private static void loadAssociationConfig() {
 
         try {
             JAXBContext jc = JAXBContext.newInstance(Foo.class);
             SAXParserFactory spf = SAXParserFactory.newInstance();
-            spf.setFeature(XMLConstants.FEATURE_SECURE_PROCESSING, true);
-            XMLReader xmlReader = spf.newSAXParser().getXMLReader();
-            InputSource inputSource = new InputSource(new FileReader("src/main/resources/input.xml"));
+            SAXParser saxParser = spf.newSAXParser();
+            SecurityManager securityManager = new SecurityManager();
+            securityManager.setEntityExpansionLimit(1);
+            saxParser.setProperty("http://apache.org/xml/properties/security-manager", securityManager);
+            XMLReader xmlReader = saxParser.getXMLReader();
+            InputSource inputSource =
+                    new InputSource(new InputStreamReader
+                            (new FileInputStream("src/main/resources/input.xml") , "UTF-8"));
             SAXSource source = new SAXSource(xmlReader, inputSource);
             Unmarshaller unmarshaller = jc.createUnmarshaller();
             Foo foo = (Foo) unmarshaller.unmarshal(source);
@@ -63,11 +69,12 @@ public class DemoWithoutSecurityManager {
             log.error("Failed to find the input.xml", e);
         } catch (ParserConfigurationException | SAXException | JAXBException e) {
             log.error("Failed to parse the input.xml", e);
+        } catch (UnsupportedEncodingException e) {
+           log.error("Encoding type is not supported");
         }
     }
 
     public static void main(String[] args) {
         loadAssociationConfig();
     }
-
 }
