+++ samples/TestEntityExpansionLimit/src/main/java/xxe/jaxb/DemoWithoutSecurityManager.java
@@ -16,47 +16,51 @@
  *  under the License.
  */
 
-package xxe;
-
+package xxe.jaxb;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
-import org.apache.xerces.util.SecurityManager;
 import org.xml.sax.InputSource;
 import org.xml.sax.SAXException;
 import org.xml.sax.XMLReader;
-
 import xxe.entity.Foo;
 
+import java.io.FileInputStream;
 import java.io.FileNotFoundException;
-import java.io.FileReader;
+import java.io.InputStreamReader;
+import java.io.UnsupportedEncodingException;
+
+import javax.xml.XMLConstants;
 import javax.xml.bind.JAXBContext;
 import javax.xml.bind.JAXBException;
 import javax.xml.bind.Unmarshaller;
 import javax.xml.parsers.ParserConfigurationException;
-import javax.xml.parsers.SAXParser;
 import javax.xml.parsers.SAXParserFactory;
 import javax.xml.transform.sax.SAXSource;
 
 
+
+
 /**
- * DemoWithSecurityManager with overiding the EntityExpansionLimit from xerces.util.SecurityManager
+ * DemoWithoutSecurityManager without overiding the EntityExpansionLimit from xerces.util.SecurityManager
+ * The default value for EntityExpansion limit from xerces.util.SecurityManager will be used
+ * even an environment property is added for the EntityExpansionLimit.
  */
 
-public class DemoWithSecurityManager {
-    private static final Log log = LogFactory.getLog(DemoWithSecurityManager.class);
+public class DemoWithoutSecurityManager {
+    private static final Log log = LogFactory.getLog(DemoWithoutSecurityManager.class);
+    private XMLReader xmlReader;
 
     private static void loadAssociationConfig() {
 
         try {
             JAXBContext jc = JAXBContext.newInstance(Foo.class);
             SAXParserFactory spf = SAXParserFactory.newInstance();
-            SAXParser saxParser = spf.newSAXParser();
-            SecurityManager securityManager = new SecurityManager();
-            securityManager.setEntityExpansionLimit(1);
-            saxParser.setProperty("http://apache.org/xml/properties/security-manager", securityManager);
-            XMLReader xmlReader = saxParser.getXMLReader();
-            InputSource inputSource = new InputSource(new FileReader("src/main/resources/input.xml"));
+            spf.setFeature(XMLConstants.FEATURE_SECURE_PROCESSING, true);
+            XMLReader xmlReader = spf.newSAXParser().getXMLReader();
+            InputSource inputSource =
+                    new InputSource(new InputStreamReader
+                            (new FileInputStream("src/main/resources/input.xml") , "UTF-8"));
             SAXSource source = new SAXSource(xmlReader, inputSource);
             Unmarshaller unmarshaller = jc.createUnmarshaller();
             Foo foo = (Foo) unmarshaller.unmarshal(source);
@@ -65,10 +69,13 @@ public class DemoWithSecurityManager {
             log.error("Failed to find the input.xml", e);
         } catch (ParserConfigurationException | SAXException | JAXBException e) {
             log.error("Failed to parse the input.xml", e);
+        } catch (UnsupportedEncodingException e) {
+            log.error("Encoding type is not supported");
         }
     }
 
     public static void main(String[] args) {
         loadAssociationConfig();
     }
+
 }
